Eternia WIP
This contract allows you to add any valid ERC20 token as a form of payment to mint your NFT, and withdraw all funds to the owner address. You can add the different currencies as you go, and even change price. It was a lot of "fun'." I learned a lot during the making, and perfecting of this contract. Much of it fell into place after watching someone implement some basic fundamentals, the "correct" way. A huge Thank you to Net2Dev for great ideas, that work.